requirements.txt
biopython
numpy
pandas
plotly